pyproject.toml
[project]
name = "powershell-chatbot-gpt-3-5-turbo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.60.2",
    "pyinstaller>=6.11.1",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
]
